Add files via upload
This program calculates the length of a real number. The programs asks for input from file "in.txt". If input is incorrect or file could not be opened or read, the programs asks user to input a new file name. The number must be between 10 and 1000 with 5 digits at most, whole and fractional parts must be separated by a comma.

This program mostly focuses on file validation.